IntegerType is now represented with a long internally

+++ src/main/java/ch/pollet/bubble/evaluation/Evaluator.java
@@ -60,6 +60,6 @@ public class Evaluator extends BubbleBaseListener {
 
     @Override
     public void exitIntegerLiteral(BubbleParser.IntegerLiteralContext ctx) {
-        evaluationContext.pushStack(new IntegerType(Integer.valueOf(ctx.getText())));
+        evaluationContext.pushStack(new IntegerType(Long.valueOf(ctx.getText())));
     }
 }
